[ADD] Creacion de servicios finalizada price es DecimalField

+++ server/services/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Service(models.Model):
     name = models.CharField('name', max_length=100)
     description = models.CharField('description', max_length=500)
-    price = models.FloatField('price', default=-1)
+    price = models.DecimalField('price', max_digits=8, decimal_places=2, default= -1.00)
     image = models.CharField('image', max_length=500)
     # reviews = models.ManyToManyField('user.User', through='Review')
